+++ app/Modules/Permissions/PermissionsModule.php
@@ -37,9 +37,9 @@ final class PermissionsModule implements Module
     {
         $router->get('/roles', [RolesController::class, 'index']);
         $router->post('/roles', [RolesController::class, 'create']);
-        $router->get('/roles/{id}', [RolesController::class, 'show']);
-        $router->post('/roles/{id}/edit', [RolesController::class, 'update']);
-        $router->post('/roles/{id}/clone', [RolesController::class, 'clone']);
-        $router->post('/roles/{id}/delete', [RolesController::class, 'delete']);
+        $router->get('/roles/{roleId}', [RolesController::class, 'show']);
+        $router->post('/roles/{roleId}/edit', [RolesController::class, 'update']);
+        $router->post('/roles/{roleId}/clone', [RolesController::class, 'clone']);
+        $router->post('/roles/{roleId}/delete', [RolesController::class, 'delete']);
     }
 }